+++ src/GitDevInsights/TrendGraph/BasicStatistics/FileExtensionChartHTMLFileGenerator.php
@@ -2,128 +2,48 @@
 
 namespace GitDevInsights\TrendGraph\BasicStatistics;
 
+use GitDevInsights\CodeInsights\Persistence\ProjectConfigDataProvider;
+use GitDevInsights\CodeInsights\Results\AnalysisResult;
 use GitDevInsights\TrendGraph\Generator\SimpleTrendGraphFilter;
+use GitDevInsights\TrendGraph\Generator\SimpleTrendGraphGenerator;
 
 class FileExtensionChartHTMLFileGenerator
 {
-    private array $jsonData;
+    private const CHART_FILENAME = 'chart-file-extension.html';
 
-    private string $chartTitle;
+    private AnalysisResult $analysisResult;
+
+    private ProjectConfigDataProvider $projectConfigDataProvider;
 
     private string $filterKey = 'language-fileext-data';
 
-    public function __construct(array $jsonData, string $chartTitle)
+    public function __construct(AnalysisResult $analysisResult, ProjectConfigDataProvider $projectConfigDataProvider)
     {
-        $this->chartTitle = $chartTitle;
-        $trendGraphFilter = new SimpleTrendGraphFilter($this->filterKey);
-        $this->jsonData = $trendGraphFilter->filterDataByValuesSet($jsonData);
+        $this->analysisResult = $analysisResult;
+        $this->projectConfigDataProvider = $projectConfigDataProvider;
     }
 
     public function renderChartOutput(): string
     {
-        // Extract dates and labels
-        $dates = array_keys(array_reverse($this->jsonData[$this->filterKey]));
-        $labels = array_keys($this->jsonData[$this->filterKey][$dates[0]]);
+        $trendGraphFilter = new SimpleTrendGraphFilter($this->filterKey);
+        $jsonData = $trendGraphFilter->filterDataByValuesSet($this->analysisResult->__toJsonData());
 
-        // Generate datasets for each label
-        $datasets = [];
-        foreach ($labels as $label) {
-            $data = [];
-            foreach ($dates as $date) {
-                $data[] = $this->jsonData[$this->filterKey][$date][$label];
-            }
+        $simpleTrendGraphGenerator = new SimpleTrendGraphGenerator(
+            $jsonData,
+            'File Extension Trend Analysis',
+            'Usage of Programming Languages by Number of Lines of Code',
+            $this->filterKey,
+            'Date',
+            'Lines of code'
+        );
 
-            $datasets[] = [
-                'label' => $label,
-                'data' => $data,
-                'fill' => false,
-                'borderColor' => 'rgb(' . rand(0, 255) . ', ' . rand(0, 255) . ', ' . rand(0, 255) . ')',
-                'tension' => 0.1,
-            ];
-        }
-
-        // Generate the HTML code
-        $html = '<!DOCTYPE html>
-<html lang="en">
-<head>
-    <style>
-        body {
-            font-family: Arial, Helvetica, sans-serif;
-        }
-        h1, h2 {
-            text-align: center;
-        }
-        h2 {
-            font-size: 13px;
-        }
-        
-        canvas {
-            border: 1px solid black;
-            margin: 0 auto;
-            width: 80vw;
-            height: 80vh;
-        }
-    </style>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>'.$this->chartTitle.' - code trends</title>
-    <!-- Include Chart.js library -->
-    <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
-</head>
-<body>
-<h1>'.$this->chartTitle.' - Programming Language Trend Analysis</h1>
-<h2>Usage of File extensions by Number of Lines of Code</h2>
-<div style="width: 80vw; height: 90vh; margin: 0 auto;">
-    <canvas id="trend_chart" width="" height=""></canvas>
-</div>
-
-<script>
-    var labels = ' . json_encode($dates) . ';
-    var datasets = ' . json_encode($datasets) . ';
-
-    // Create the chart
-    var ctx = document.getElementById("trend_chart").getContext("2d");
-    var myChart = new Chart(ctx, {
-        type: "line",
-        data: {
-            labels: labels,
-            datasets: datasets,
-        },
-        options: {
-            scales: {
-                x: {
-                    type: "category",
-                    title: {
-                        display: true,
-                        text: "Date"
-                    }
-                },
-                y: {
-                    beginAtZero: true,
-                    title: {
-                        display: true,
-                        text: "Number of Lines of Code"
-                    }
-                }
-            },
-            plugins: {
-                legend: {
-                    display: true,
-                    position: "bottom"
-                }
-            }
-        }
-    });
-</script>
-</body>
-</html>';
-
-        return $html;
+        return $simpleTrendGraphGenerator->renderChartOutput();
     }
 
-    public function writeChartOutputToFile(string $fileName) : void
+    public function writeChartOutputToFile() : void
     {
         $html = $this->renderChartOutput();
+        $fileName = $this->projectConfigDataProvider->analyseResultPath.'/'.self::CHART_FILENAME;
         file_put_contents($fileName, $html);
     }
 }
